requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Django==4.2.16
django-allauth==65.3.0
django-bootstrap-v5==1.0.11
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
webcolors==24.11.1
